+++ Cefalo.TechDaily.Api/Controllers/UserController.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Mvc;
 using System.Security.Cryptography;
 using Cefalo.TechDaily.Service.Dto;
 using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 
 namespace Cefalo.TechDaily.Api.Controllers
 {
@@ -15,9 +16,11 @@ namespace Cefalo.TechDaily.Api.Controllers
     public class UserController : ControllerBase
     {
         private readonly IUserService _userService;
-        public UserController(IUserService userService)
+        private readonly IAuthService _authService;
+        public UserController(IUserService userService, IAuthService authService)
         {
             _userService = userService;
+            _authService = authService;
         }
         [HttpGet]
         public async Task<ActionResult<IEnumerable<UserDto?>>> Getusers()
@@ -38,18 +41,22 @@ namespace Cefalo.TechDaily.Api.Controllers
             if (userDto == null) return BadRequest("Cant create user");
             return CreatedAtAction(nameof(PostUser),userDto);
         }
-        [HttpPatch("{Username}")]
+        [HttpPatch("{Username}"), Authorize]
         public async Task<IActionResult> UpdateUser(string Username, UpdateUserDto updateUserDto)
         {
             //if (Username != updateUserDto.Username) return BadRequest("Username does not match");
             //System.Diagnostics.Debug.WriteLine("hello");
+            var loggedInUser = _authService.GetMyName();
+            if (loggedInUser != Username) return BadRequest("You are not authorized to update this user");
             var userDto = await _userService.UpdateUser(Username,updateUserDto);
             if (userDto == null) return BadRequest("User not found");
             return Ok(userDto);
         }
-        [HttpDelete("{Username}")]
+        [HttpDelete("{Username}"), Authorize]
         public async Task<IActionResult> DeleteUser(string Username)
         {
+            var loggedInUser = _authService.GetMyName();
+            if (loggedInUser != Username) return BadRequest("You are not authorized to delete this user");
             var deleted = await _userService.DeleteUser(Username);
             if(!deleted) return BadRequest("User not found");
             return NoContent();
